Move consumption budget helpers out of azure/helpers
