+++ azurerm/internal/services/consumption/consumption_budget_subscription_resource.go
@@ -2,6 +2,7 @@ package consumption
 
 import (
 	"fmt"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/consumption/parse"
 	"time"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
@@ -25,7 +26,7 @@ func resourceArmConsumptionBudgetSubscription() *schema.Resource {
 			Delete: schema.DefaultTimeout(30 * time.Minute),
 		},
 
-		Schema: azure.SchemaAzureConsumptionBudgetSubscriptionResource(),
+		Schema: SchemaAzureConsumptionBudgetSubscriptionResource(),
 	}
 }
 
@@ -48,7 +49,7 @@ func resourceArmConsumptionBudgetSubscriptionRead(d *schema.ResourceData, meta i
 	}
 
 	// Parse the consumption budget to get the scope
-	consumptionBudgetId, err := azure.ParseAzureConsumptionBudgetID(d.Id())
+	consumptionBudgetId, err := parse.ConsumptionBudgetID(d.Id())
 	if err != nil {
 		return err
 	}
