+++ azurerm/internal/services/consumption/validate/name.go
@@ -1,0 +1,14 @@
+package validate
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
+	"regexp"
+)
+
+func ConsumptionBudgetName() schema.SchemaValidateFunc {
+	return validation.StringMatch(
+		regexp.MustCompile("^[-_a-zA-Z0-9]{1,63}$"),
+		"The consumption budget name can contain only letters, numbers, underscores, and hyphens. The consumption budget name be between 6 and 63 characters long.",
+	)
+}
